+++ docs/changelog.rst
@@ -6,6 +6,7 @@ Changes in 0.6.X
 ================
 
 - Fixed BinaryField python value issue (MongoEngine/mongoengine#48)
+- Fixed SequenceField non numeric value lookup (MongoEngine/mongoengine#41)
 - Fixed queryset manager issue (MongoEngine/mongoengine#52)
 - Fixed FileField comparision (hmarr/mongoengine#547)